[trade] arrage files.
