+++ src/test/java/com/itchain/samplemsa/samplemsa/TestApplication.java
@@ -5,9 +5,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MockApplication {
+public class TestApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(MockApplication.class, args);
+        SpringApplication.run(TestApplication.class, args);
     }
 }
