+++ src/test/java/com/itchain/samplemsa/samplemsa/TestApplicationTests.java
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class MockApplicationTests {
+public class TestApplicationTests {
     @Test
     public void contextLoads() {
     }
